requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
Brotli==1.1.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
ddgs==9.6.1
fastapi==0.120.1
filelock==3.20.0
fsspec==2025.9.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.0.1
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
packaging==25.0
pillow==12.0.0
primp==0.15.0
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
shellingham==1.5.4
smolagents==1.22.0
sniffio==1.3.1
socksio==1.0.0
starlette==0.49.1
tqdm==4.67.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
